# inspection/upper_computer/frontend/openapi/inspection_gateway_openapi.json
{
  "components": {
    "schemas": {
      "ChangeGatewayPasswordRequest": {
        "additionalProperties": false,
        "description": "Change the password of the current authenticated session.",
        "properties": {
          "currentPassword": {
            "description": "Current password bound to the active session.",
            "minLength": 1,
            "title": "Currentpassword",
            "type": "string"
          },
          "newPassword": {
            "description": "Replacement password for the active session.",
            "minLength": 1,
            "title": "Newpassword",
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "title": "ChangeGatewayPasswordRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginGatewaySessionRequest": {
        "additionalProperties": false,
        "description": "Authenticate one gateway session.\n\nArgs:\n    username: Operator login name.\n    password: Operator password.\n\nReturns:\n    The request payload serialized through :meth:`to_payload`.\n\nRaises:\n    ValidationError: Raised by Pydantic when required fields are missing or\n        empty.\n\nBoundary behavior:\n    Unknown JSON keys are rejected so the login contract cannot silently\n    drift from the frontend SDK.",
        "properties": {
          "password": {
            "description": "Operator password.",
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "description": "Operator login name.",
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginGatewaySessionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Inspection HMI Gateway",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/change-password": {
      "post": {
        "operationId": "changeGatewayPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeGatewayPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Changegatewaypassword",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "loginGatewaySession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginGatewaySessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Logingatewaysession",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logoutGatewaySession",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Logoutgatewaysession",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "operationId": "getGatewaySession",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Getgatewaysession",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Session Info",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/ws-ticket": {
      "post": {
        "operationId": "issueGatewayWsTicket",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Issuegatewaywsticket",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Issue Ws Ticket",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/results": {
      "get": {
        "operationId": "getInspectionResults",
        "parameters": [
          {
            "in": "query",
            "name": "batchId",
            "required": false,
            "schema": {
              "default": "",
              "title": "Batchid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "recipeId",
            "required": false,
            "schema": {
              "default": "",
              "title": "Recipeid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "decision",
            "required": false,
            "schema": {
              "default": "",
              "title": "Decision",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "defectType",
            "required": false,
            "schema": {
              "default": "",
              "title": "Defecttype",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "qrText",
            "required": false,
            "schema": {
              "default": "",
              "title": "Qrtext",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "default": "",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "default": "",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Getinspectionresults",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Results",
        "tags": [
          "results"
        ]
      }
    },
    "/api/v1/results/read-model/repair": {
      "post": {
        "operationId": "repairInspectionReadModel",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Repairinspectionreadmodel",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Repair Read Model",
        "tags": [
          "results"
        ]
      }
    },
    "/api/v1/results/read-model/status": {
      "get": {
        "operationId": "getInspectionReadModelStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Getinspectionreadmodelstatus",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Read Model Status",
        "tags": [
          "results"
        ]
      }
    },
    "/api/v1/results/statistics": {
      "get": {
        "operationId": "getInspectionResultStatistics",
        "parameters": [
          {
            "in": "query",
            "name": "batchId",
            "required": false,
            "schema": {
              "default": "",
              "title": "Batchid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "recipeId",
            "required": false,
            "schema": {
              "default": "",
              "title": "Recipeid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "decision",
            "required": false,
            "schema": {
              "default": "",
              "title": "Decision",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "defectType",
            "required": false,
            "schema": {
              "default": "",
              "title": "Defecttype",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "qrText",
            "required": false,
            "schema": {
              "default": "",
              "title": "Qrtext",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "default": "",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "default": "",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sampleLimit",
            "required": false,
            "schema": {
              "default": 120,
              "maximum": 1000,
              "minimum": 1,
              "title": "Samplelimit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Getinspectionresultstatistics",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Result Statistics",
        "tags": [
          "results"
        ]
      }
    },
    "/api/v1/results/summary/{batch_id}": {
      "get": {
        "operationId": "getBatchResultSummary",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Getbatchresultsummary",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Result Summary",
        "tags": [
          "results"
        ]
      }
    },
    "/api/v1/results/{result_id}": {
      "get": {
        "operationId": "getInspectionResultDetail",
        "parameters": [
          {
            "in": "path",
            "name": "result_id",
            "required": true,
            "schema": {
              "title": "Result Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Getinspectionresultdetail",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Result Detail",
        "tags": [
          "results"
        ]
      }
    },
...
  }
}